requirements.txt
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
numpy>=1.24.0
open3d>=0.19.0
matplotlib>=3.7.0
requests>=2.31.0
trimesh>=4.0.0
scipy>=1.10.0
rembg>=2.0.0
onnxruntime>=1.15.0
opencv-python>=4.7.0
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
numpy>=1.24.0
open3d>=0.19.0
matplotlib>=3.7.0
requests>=2.31.0
trimesh>=4.0.0
scipy>=1.10.0
rembg>=2.0.0
onnxruntime>=1.15.0
opencv-python>=4.7.0
timm>=0.9.0
git+https://github.com/openai/shap-e.git